requirements.txt
antiorm==1.2.1
db==0.1.1
db-sqlite3==0.0.1
gmqtt==0.7.0
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
